Change length variable in target_string_use from size_t to int and cast return from strlen. Archive the binaries for Windows and push.

+++ g_misc.c
@@ -1584,10 +1584,10 @@ void target_string_use (edict_t *self, edict_t *other, edict_t *activator)
 {
 	edict_t *e;
 	int		n;
-	size_t	l;
+	int		l;
 	char	c;
 
-	l = strlen(self->message);
+	l = (int)strlen(self->message);
 	for (e = self->teammaster; e; e = e->teamchain)
 	{
 		if (!e->count)
